// src/generated/data-contracts.ts
export enum CurrencyEnum {
  USD = 'USD',
  ZAR = 'ZAR',
  AED = 'AED',
}
// ...
export enum OrderByEnum {
  Asc = 'asc',
  Desc = 'desc',
}
// ...
export type LoanResponse = (NewLoan & object) & {
  loanOfficerUser?: LoanOfficerUser;
  /**
   * Unique loan identifier
   * @example "BtCSNcteA2oD226"
   */
  loanId?: string;
  /**
   * Loan current status
   * @example "Pending"
   */
  loanStatus?: LoanResponseLoanStatusEnum;
  /**
   * The date finance team transfers the funds to the client
   * @example "2023-05-06T00:00:00.000Z"
   */
  disbursementDate?: string;
  /**
   * The date on which the loan is due for final repayment
   * @example "2024-05-06T00:00:00.000Z"
   */
  maturityDate?: string;
  /** Currency */
  currency?: CurrencyEnum;
  /**
   * Date of which the loan is  actually paid off
   * @example "2024-12-06T00:00:00.000Z"
   */
  payoffDate?: string;
  /**
   * Date the loan was approved
   * @example "2023-12-06T00:00:00.000Z"
   */
  loanApprovalDate?: string;
  /**
   * How many times the loan has been deliquent
   * @example 0
   */
  timesDelinquent?: number;
  /**
   * The principal left to be paid
   * @example 0
   */
  netPayablePrincipal?: number;
  /**
   * Late payment fees for this loan
   * @example 0
   */
  netPayablePenalties?: number;
  /**
   * Total left sum of pricnipal, interest, penalties amount
   * @example 0
   */
  netPayableTotal?: number;
  /**
   * Total left sum of interest amount
   * @example 0
   */
  netPayableInterest?: number;
  /**
   * Maximum allowable disbursements including offset
   * @example 0
   */
  maxAllowableDisbursements?: number;
  /**
   * The total money disbursed to the customer
   * @example 0
   */
  totalDisbursedAmount?: number;
  /**
   * Total Principal paid by the customer
   * @example 0
   */
  paidPrincipalAmount?: number;
  /**
   * Total interest paid by the customer
   * @example 0
   */
  paidInterestAmount?: number;
  /**
   * Total penalties paid by the customer
   * @example 0
   */
  paidPenaltiesAmount?: number;
  /**
   * Late payment fees for this loan
   * @example 0
   */
  totalPenaltyAmount?: number;
  /**
   * Total offsets for the loan
   * @example 0
   */
  totalOffsetAmount?: number;
  /**
   * Total amount waived off
   * @example 0
   */
  waivedAmount?: number;
  /**
   * Penalty amount waived off
   * @example 0
   */
  waivedPenaltyAmount?: number;
  /**
   * Total interest adjusted
   * @example 0
   */
  adjustedInterestAmount?: number;
  /**
   * Total interest discounted
   * @example 0
   */
  allocatedDiscountInterestAmount?: number;
  /**
   * Principal Amount Written off; this is the amount considered not recoverable
   * @example 0
   */
  writtenOffPrincipalAmount?: number;
  /**
   * Interests Amount Written off; this is the amount considered not recoverable
   * @example 0
   */
  writtenOffInterestAmount?: number;
  /**
   * Total Amount Written off; this is the amount considered not recoverable
   * @example 0
   */
  totalWrittenOffAmount?: number;
  /**
   * This is the amount which was  restructured as part of a different loan
   * @example 0
   */
  restructuredAmount?: number;
  /**
   * The amount of interest
   * @example 1500
   */
  interestAmount?: number;
  /**
   * Date of next payment
   * @example "2023-05-06T00:00:00.000Z"
   */
  nextPaymentDate?: string;
  /**
   * Number of payments equal to 1 month
   * @example 12
   */
  numberOfPayments?: number;
  /**
   * Number of current payment
   * @example 0
   */
  currentPaymentNumber?: number;
  /**
   * Current delinquent amount of loan
   * @example 0
   */
  delinquentAmount?: number;
  /**
   * Origination fee
   * @example 0
   */
  originationFee?: number;
  /**
   * Origination fee amount
   * @example 0
   */
  originationFeeAmount?: number;
  /**
   * Origination fee amount
   * @example 0
   */
  totalChargesAmount?: number;
};

export interface DisbursementRequestResponse {
  /**
   * Disbursement Request ID.
   * @example "DRMYcrC9orZztiwQp"
   */
  disbursementRequestId?: string;
  /**
   * Loan ID to which Disbursement Request refers.
   * @example "DRMYcrC9orZztiwQp"
   */
  loanId?: string;
  /**
   * Loan Name
   * @example "My Company Inc EG"
   */
  loanName?: string;
  /**
   * Company Identifier
   * @example 123456
   */
  companyIdentifier?: string;
  /**
   * Company Legal Name
   * @example "My Company Inc"
   */
  companyLegalName?: string;
  /**
   * Salesforce Account ID
   * @example 123456
   */
  salesforceAccountId?: string;
  /**
   * Requested disbursement amount
   * @example "1000.0"
   */
  amount?: string;
  /** Currency */
  currency?: CurrencyEnum;
  /**
   * Requested disbursement date
   * @example "2024-08-29"
   */
  paymentDate?: string;
  /**
   * Datetime when request has been requested
   * @example "2024-07-01 12:00:00"
   */
  requestedOn?: string;
  requestedBy?: LoanOfficerUser;
  /**
   * Disbursement Request Status
   * @example "NEW"
   */
  status?: DisbursementRequestResponseStatusEnum;
  /**
   * Disbursement Request Status
   * @example "Normal"
   */
  priority?: DisbursementRequestResponsePriorityEnum;
  /**
   * Datetime when request has been approved
   * @example "2024-07-01 12:00:00"
   */
  approvedOn?: string;
  approvedBy?: LoanOfficerUser;
  /**
   * Datetime when request has been canceled
   * @example "2024-07-01 12:00:00"
   */
  canceledOn?: string;
  canceledBy?: LoanOfficerUser;
  _links?: HalLink[];
}

export interface DisbursementBulkApproval {
  /**
   * TOTP Code
   * @example "200200"
   */
  totp?: string;
  /**
   * List of disbursement request ids to be approved
   * @example ["disbursementId1","disbursementId2"]
   */
  disbursementRequestIds?: any[];
}

export interface BulkDisbursementRequestApproveResponse {
  disbursementRequestResults?: (DisbursementRequestResponse & {
    isSuccess?: boolean;
    error?: string;
  })[];
}

export type NewDisbursementRequestResponse = (DisbursementRequestResponse & object) & {
  _embedded?: {
    loan?: LoanResponse & object;
    _links?: HalLink[];
  };
};
// ...
export interface RevioEftPayment {
  /** Created At */
  created_at?: string;
  /** Updated At */
  updated_at?: string;
  /** Payment Id */
  payment_id: string;
  /** Reference Number */
  reference_number: string | null;
  /** External Reference */
  external_reference: string | null;
  /** Payment Date */
  payment_date: string | null;
  /** State */
  state: string | null;
  /** Amount Cents */
  amount_cents: number | null;
  /** Currency */
  currency: string | null;
  /** Client Id */
  client_id: string | null;
  /** Eft Payment Data */
  eft_payment_data?: object | null;
}
// ...
export type HalLink = Record<
  string,
  {
    href?: string;
  }
>;
// ...
export interface LoanOfficerUser {
  /** @example "123e4567-e89b-12d3-a456-426614174000" */
  user_id?: string;
  /** @example "example@example.com" */
  email?: string;
}
// ...
export enum DisbursementRequestResponseStatusEnum {
  NEW = 'NEW',
  APPROVED = 'APPROVED',
  PAID = 'PAID',
  FAILED = 'FAILED',
  CANCELED = 'CANCELED',
}

/**
 * Disbursement Request Status
 * @example "Normal"
 */
export enum DisbursementRequestResponsePriorityEnum {
  Normal = 'Normal',
  Urgent = 'Urgent',
}
// ...
export interface GetDisbursementRequestsParams {
  /** In which order will the data be sorted (ascending or descending) */
  orderBy?: OrderByEnum;
  /** Search items by partial matching of search param */
  search?: string;
  /** Number of items in the response */
  limit?: number;
  /** Page number */
  page?: number;
}

export interface GetDisbursementRequestsData {
  disbursementRequests?: DisbursementRequestResponse[];
}

export type PostApproveBulkDisbursementRequestsData = BulkDisbursementRequestApproveResponse;

export type ResetRiskFlagData = (LoanResponse & object) & {
  _links?: HalLink[];
};

export type PostCancelDisbursementRequestsData = NewDisbursementRequestResponse;

export type PostApproveDisbursementRequestsData = NewDisbursementRequestResponse;

export type DisbursementRequestRevioStatusUpdateData = NewDisbursementRequestResponse;